Simplify cache codes
